Kamal (#41)
* #1

* complete AddToken Feature

* AddToken Completed

* Posts

* changes in lead

* sales updated

* 20/08/2025

* 02/09/2025

+++ src/main/java/com/realestate/main/repository/ReviewsRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.realestate.main.entity.Reviews;
 
 @Repository
-public interface ReviewsRepository extends JpaRepository<Reviews, Long> {
+public interface ReviewsRepository extends JpaRepository<Reviews, Integer> {
 	List<Reviews> findByAgentEmail(String email);
 }
